Split out hotkey settings in its own tab
