+++ LightBulb/Views/Components/HotKeySettingsTabView.xaml.cs
@@ -1,0 +1,10 @@
+﻿namespace LightBulb.Views.Components
+{
+    public partial class HotKeySettingsTabView
+    {
+        public HotKeySettingsTabView()
+        {
+            InitializeComponent();
+        }
+    }
+}
